Moved image resizing to seperate function to make sure all images are resized only once, more fully definied what resolutions/aspect ratios are valid. Moving to rerun texton dict and other tests, outline thesis, list questions for later

+++ dictCreation.cpp
@@ -65,7 +65,7 @@ vector<float> createBins(Mat texDic){
   return v;
 }
 
-void dictCreateHandler(int cropsize, double scale, int numClusters){
+void dictCreateHandler(int cropsize, int scale, int numClusters){
   int dictSize = numClusters;
   int attempts = 5;
   int flags = KMEANS_PP_CENTERS;
