+++ dictCreation.h
@@ -6,6 +6,6 @@ void binLimits(std::vector<float>& tex);
 void removeDups(std::vector<float>& v);
 std::vector<float> matToVec(cv::Mat m);
 std::vector<float> createBins(cv::Mat texDic);
-void dictCreateHandler(int cropsize, double scale, int numClusters);
+void dictCreateHandler(int cropsize, int scale, int numClusters);
 
 #endif
