+++ modelBuild.h
@@ -2,6 +2,6 @@
 #ifndef modelBuild
 #define modelBuild
 
-void modelBuildHandle(int cropsize, double scale, int numClusters);
+void modelBuildHandle(int cropsize, int scale, int numClusters);
 
 #endif
